refactor: add detail packages in models
