+++ src/main/java/coztion/springai/core/vertex/application/mapper/VertexCommandMapper.java
@@ -1,9 +1,9 @@
 package coztion.springai.core.vertex.application.mapper;
 
-import coztion.springai.core.vertex.application.model.VertexClaudeGenerateCommand;
-import coztion.springai.core.vertex.application.model.VertexGeminiGenerateCommand;
-import coztion.springai.core.vertex.application.model.VertexGenerateCommand;
-import coztion.springai.core.vertex.application.model.VertexImagenGenerateCommand;
+import coztion.springai.core.vertex.application.model.command.VertexClaudeGenerateCommand;
+import coztion.springai.core.vertex.application.model.command.VertexGeminiGenerateCommand;
+import coztion.springai.core.vertex.application.model.command.VertexGenerateCommand;
+import coztion.springai.core.vertex.application.model.command.VertexImagenGenerateCommand;
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
