+++ src/main/java/coztion/springai/core/vertex/application/model/command/VertexGeminiGenerateCommand.java
@@ -1,4 +1,4 @@
-package coztion.springai.core.vertex.application.model;
+package coztion.springai.core.vertex.application.model.command;
 
 import com.google.genai.types.Content;
 import com.google.genai.types.GenerateContentConfig;
