+++ src/main/java/coztion/springai/core/vertex/application/model/result/VertexImagenGenerateResult.java
@@ -1,4 +1,4 @@
-package coztion.springai.core.vertex.application.model;
+package coztion.springai.core.vertex.application.model.result;
 
 import java.util.List;
 import lombok.AccessLevel;
