+++ src/main/java/coztion/springai/core/vertex/application/service/VertexClaudeService.java
@@ -8,8 +8,8 @@ import com.anthropic.models.messages.RawMessageStreamEvent;
 import com.anthropic.models.messages.TextDelta;
 import com.anthropic.vertex.backends.VertexBackend;
 import com.google.auth.oauth2.GoogleCredentials;
-import coztion.springai.core.vertex.application.model.VertexClaudeGenerateCommand;
-import coztion.springai.core.vertex.application.model.VertexClaudeGenerateResult;
+import coztion.springai.core.vertex.application.model.command.VertexClaudeGenerateCommand;
+import coztion.springai.core.vertex.application.model.result.VertexClaudeGenerateResult;
 import java.io.IOException;
 import java.util.Iterator;
 import org.springframework.beans.factory.annotation.Value;
