+++ src/main/java/coztion/springai/core/vertex/application/service/VertexGeminiService.java
@@ -6,8 +6,8 @@ import com.google.genai.ResponseStream;
 import com.google.genai.types.Content;
 import com.google.genai.types.GenerateContentConfig;
 import com.google.genai.types.GenerateContentResponse;
-import coztion.springai.core.vertex.application.model.VertexGeminiGenerateCommand;
-import coztion.springai.core.vertex.application.model.VertexGeminiGenerateResult;
+import coztion.springai.core.vertex.application.model.command.VertexGeminiGenerateCommand;
+import coztion.springai.core.vertex.application.model.result.VertexGeminiGenerateResult;
 import java.io.IOException;
 import java.util.Iterator;
 import lombok.SneakyThrows;
