+++ src/main/java/coztion/springai/core/vertex/application/service/VertexImagenService.java
@@ -3,8 +3,8 @@ package coztion.springai.core.vertex.application.service;
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.genai.Client;
 import com.google.genai.types.*;
-import coztion.springai.core.vertex.application.model.VertexImagenGenerateCommand;
-import coztion.springai.core.vertex.application.model.VertexImagenGenerateResult;
+import coztion.springai.core.vertex.application.model.command.VertexImagenGenerateCommand;
+import coztion.springai.core.vertex.application.model.result.VertexImagenGenerateResult;
 import java.io.IOException;
 import java.util.*;
 import org.springframework.beans.factory.annotation.Value;
